+++ hpl/src/pauxil/HPL_dlaswp03N.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -146,7 +146,8 @@ void HPL_dlaswp03N
    {
       for( i = 0; i < M; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*( W0 + (size_t)(i) * (size_t)(LDW) )); 
+         w0 = w + (size_t)(i) * (size_t)(LDW);
 
          *u0 = w0[ 0]; u0 += LDU;
 #if ( HPL_LASWP03N_DEPTH >  1 )
@@ -182,7 +183,8 @@ void HPL_dlaswp03N
    {
       for( i = 0; i < M; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*( W0 + (size_t)(i) * (size_t)(LDW) )); 
+         w0 = w + (size_t)(i) * (size_t)(LDW);
          for( j = 0; j < nr; j++, u0 += LDU ) { *u0 = w0[j]; }
       }
    }
