+++ hpl/src/pauxil/HPL_dlaswp04N.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -179,7 +179,9 @@ void HPL_dlaswp04N
    {
       for( i =  0; i < M0; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U + LINDXAU[i]; w0 = w + i * LDW;
+         a0 = A + (size_t)(LINDXA[i]);
+         u0 = U + (size_t)(LINDXAU[i]);
+         w0 = w + (size_t)(i) * (size_t)(LDW);
 
          *a0 = *u0; *u0 = w0[ 0]; a0 += LDA; u0 += LDU;
 #if ( HPL_LASWP04N_DEPTH >  1 )
@@ -227,7 +229,8 @@ void HPL_dlaswp04N
 
       for( i = M0; i < M1; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*(W0+(size_t)(i)*(size_t)(LDW)));
+         w0 = w + (size_t)(i) * (size_t)(LDW);
 
          *u0 = w0[ 0]; u0 += LDU;
 #if ( HPL_LASWP04N_DEPTH >  1 )
@@ -263,13 +266,16 @@ void HPL_dlaswp04N
    {
       for( i = 0; i < M0; i++ )
       {
-         a0 = A + LINDXA[i]; u0 = U + LINDXAU[i]; w0 = w + i * LDW;
+         a0 = A + (size_t)(LINDXA[i]);
+         u0 = U + (size_t)(LINDXAU[i]);
+         w0 = w + (size_t)(i) * (size_t)(LDW);
          for( j = 0; j < nr; j++, a0 += LDA, u0 += LDU )
          { *a0 = *u0; *u0 = w0[j]; }
       }
       for( i = M0; i < M1; i++ )
       {
-         u0 = U + (int)(*(W0+i*LDW)); w0 = w + i * LDW;
+         u0 = U + (size_t)(*(W0+(size_t)(i)*(size_t)(LDW)));
+         w0 = w + (size_t)(i) * (size_t)(LDW);
          for( j = 0; j < nr; j++, u0 += LDU ) { *u0 = w0[j]; }
       }
    }
