+++ hpl/src/pgesv/HPL_pdlaswp00T.c
@@ -1,10 +1,10 @@
 /* 
  * -- High Performance Computing Linpack Benchmark (HPL)                
- *    HPL - 1.0a - January 20, 2004                          
+ *    HPL - 2.0 - September 10, 2008                          
  *    Antoine P. Petitet                                                
  *    University of Tennessee, Knoxville                                
- *    Innovative Computing Laboratories                                 
- *    (C) Copyright 2000-2004 All Rights Reserved                       
+ *    Innovative Computing Laboratory                                 
+ *    (C) Copyright 2000-2008 All Rights Reserved                       
  *                                                                      
  * -- Copyright notice and Licensing terms:                             
  *                                                                      
@@ -22,7 +22,7 @@
  * 3. All  advertising  materials  mentioning  features  or  use of this
  * software must display the following acknowledgement:                 
  * This  product  includes  software  developed  at  the  University  of
- * Tennessee, Knoxville, Innovative Computing Laboratories.             
+ * Tennessee, Knoxville, Innovative Computing Laboratory.             
  *                                                                      
  * 4. The name of the  University,  the name of the  Laboratory,  or the
  * names  of  its  contributors  may  not  be used to endorse or promote
@@ -149,7 +149,9 @@ void HPL_pdlaswp00T
 /*
  * Allocate space for temporary W (ldW * jb)
  */
-   vptr = (void*)malloc( (align + (jb * ldW)) * sizeof(double) );
+   vptr = (void*)malloc( ( (size_t)(align) + 
+                           ((size_t)(jb) * (size_t)(ldW))) * 
+                           sizeof(double) );
    if( vptr == NULL )
    { HPL_pabort( __LINE__, "HPL_pdlaswp00T", "Memory allocation failed" ); }
